+++ Sources/KeyboardKit/Layout/KeyboardLayoutConfiguration.swift
@@ -58,7 +58,7 @@ public extension KeyboardLayoutConfiguration {
      */
     static func standard(
         for context: KeyboardContext) -> KeyboardLayoutConfiguration {
-        #if os(iOS) || os(watchOS) || os(macOS)
+        #if os(iOS) || os(macOS) || os(watchOS)
         standard(
             forIdiom: context.device.userInterfaceIdiom,
             screenSize: context.screen.bounds.size,
@@ -68,7 +68,7 @@ public extension KeyboardLayoutConfiguration {
         #endif
     }
     
-    #if os(iOS) || os(watchOS) || os(macOS)
+    #if os(iOS) || os(macOS) || os(watchOS)
     /**
      The standard config for the provided device and screen.
      */