+++ Sources/KeyboardKit/_Deprecated/SystemAudioPlayer+Deprecated.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public extension SystemAudioPlayer {
+    
+    @available(*, deprecated, renamed: "play")
+    func playSystemAudio(_ audio: SystemAudio) {
+        play(audio)
+    }
+}